requirements.txt
asgiref==3.6.0 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
coverage==7.2.2
django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1 ; python_version < '3.11'
factory-boy==3.2.1
faker==17.6.0 ; python_version >= '3.7'
iniconfig==2.0.0 ; python_version >= '3.7'
packaging==23.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pyjwt==2.6.0 ; python_version >= '3.7'
pytest==7.2.2 ; python_version >= '3.7'
pytest-django==4.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
tomli==2.0.1 ; python_version < '3.11'
